requirements.txt
alembic==1.8.1
Bootstrap-Flask==2.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
DateTime==4.7
distlib==0.3.6
filelock==3.4.2
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.0.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
packaging==21.3
pipenv==2023.2.18
platformdirs==2.4.1
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.43
urllib3==1.26.15
virtualenv==20.19.0
virtualenv-clone==0.5.7
Werkzeug==2.2.2
WTForms==3.0.1
zope.interface==5.5.2
